+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [10.6.0] - 2024.01.10
+### Added
+- Api key authentication strategy added
+- X-WSSE authentication strategy added
+
 ## [10.5.0] - 2023.12.22
 ### Added
 - `oneOf` schema support